requirements.txt
openai-whisper
gradio
torch
transformers
sentence-transformers
ffmpeg-python
